backend/requirements.txt
fastapi
uvicorn
supabase
anthropic
python-jobspy
python-docx
markdownify
python-multipart
python-dotenv
httpx
